Add reply support to board functionality
This commit introduces the ability for replies to be made on a board. Reply.java is added as a new domain class and relationships are set with Board.java, while Board.java is also updated to account for replies. ReplyRepository is also set up along with custom implemented methods to facilitate the necessary operations with replies. ReplyService is provided to handle reply-related logic and ReplyController interfaces it with the outside. Changes also extend to updating BoardService to work with replies, modifying request and responses to fit the new design, and other minor adjustments across board-related components.