Implement rate limiting and refactor PricingPlan enum
The codebase has been updated to implement rate limiting on specific endpoints. This was accomplished by annotating the desired controllers with the 'RateLimited' annotation which sets a limit on the number of times an endpoint can be hit.

In addition to this, the PricingPlan enum was refactored to adjust the number of requests allowed within a certain timeframe. For example, the free tier now refreshes 10 tokens every minute. The ApiKey class has also been updated to use ordinal enumeration, which could potentially optimize querying efficiency.

Furthermore, some unnecessary imports were removed to clean up code, and CORS settings were updated to allow specific methods and headers. An error log was also added in the SecurityResourceService for better error monitoring and debugging purpose.