Update S3 file copy command in build workflow
The S3 file copy command in the build workflow has been updated with the full path of "yeoboya-lunch-api-0.0.1-SNAPSHOT.jar". This ensures that the file is found in the right location before being uploaded to S3, improving script reliability and accuracy. Change in 'Sync output files to S3'-command enhances the debugging process.