Update exception handling and add token ignore url feature
The commit refines the logging of exceptions thrown in the Jwt token provider by removing the exception stack trace from the log messages. It also adds a new feature to ignore certain URLs from token checks. A new database table is added to keep track of URLs to be ignored, and the logic of the Jwt authentication filter is updated to consult this table before checking tokens. A new repository and domain class are created for database interaction. Some boilerplate configurations in the security config and application prod files are also updated.