Inject Environment bean into RootController and update API doc link
The Environment bean was injected into the RootController java file. The "api-documentation" entry in getRootInfo() has been modified to dynamically switch the documentation link based on the active profile. When the active profile is "prod", the documentation link will point toward the production site. Otherwise, it will point to localhost.