Refactor MemberProfile retrieval for MemberService
The MemberController and MemberService classes have been updated to fetch, process and return profile image files for a member. The MemberResponse is now equipped with a new List attribute to carry multiple profile image paths. This change enables the users of our API to retrieve multiple profile images for a single member.