Added BoardEdit functionality in BoardService and BoardController
Added a new feature that allows the editing of existing board posts. A new class, BoardEdit, was created to handle the edit requests. The function 'editBoard' was added to the 'BoardService' class and correspondingly, a PATCH mapping was added to the 'BoardController' class to handle the corresponding board editing requests.