Add OAuth configuration for Google and Naver in prod-environment
This commit introduces a new YAML configuration file named 'application-prod-oauth.yml'. It seconds the registration and provider configurations for Google and Naver in a production environment. This file also includes client secrets and ids along with the necessary redirect URIs. For improved security, the client secrets and id's have been encrypted.